+++ test/fixtures/ImplementationFixture.sol
@@ -61,7 +61,7 @@ abstract contract ImplementationFixture is BaseConstants, Harness {
         MockBaseModule module_,
         uint32 moduleId_,
         uint16 moduleType_,
-        uint16 moduleVersion_,
+        uint32 moduleVersion_,
         bool moduleUpgradeable_,
         bool moduleRemoveable_
     ) internal {
